requirements.txt
Flask==3.0.3
gunicorn==22.0.0
pdfminer.six==20250506
Werkzeug==3.0.4
